Enlarged an INTEGER to BIGINT
Former-commit-id: 7fd9e0715dcb3d93ce6a7dd5f96cbee4d32a1c3c

+++ openquake/engine/db/schema/openquake.sql
@@ -191,8 +191,8 @@ CREATE TABLE uiapi.performance (
     task VARCHAR,
     operation VARCHAR NOT NULL,
     duration FLOAT,
-    pymemory INTEGER,
-    pgmemory INTEGER
+    pymemory BIGINT,
+    pgmemory BIGINT
 )  TABLESPACE uiapi_ts;
